move window config code to a separate file
